mobile: moves schema methods to own file
